update dependecies and refactoring
