+++ pyproject.toml
@@ -6,7 +6,7 @@ requires = ["setuptools>=40.8.0", "wheel"]
 name = "reduct-cli"
 version = "0.2.0"
 description = "CLI client for Reduct Storage"
-requires-python = ">=3.8"
+requires-python = ">=3.7"
 readme = "README.md"
 license = { file = "LICENSE" }
 keywords = ["cli", "reduct-storage", "tool", "database", "time series database"]
@@ -20,7 +20,7 @@ maintainers = [
 ]
 
 classifiers = [
-    "Development Status :: 3 - Alpha",
+    "Development Status :: 4 - Beta",
     "Programming Language :: Python :: 3 :: Only",
     "Environment :: Console",
     "Intended Audience :: Information Technology",
@@ -31,7 +31,7 @@ classifiers = [
 ]
 
 dependencies = [
-    "reduct-py~=1.0",
+    "reduct-py~=1.1",
     "click~=8.1",
     "tomlkit~=0.11",
     "rich~=12.6",
