+++ reduct_cli/utils/consoles.py
